split classes into modules
